+++ MyFood/Models/SafetyTool.cs
@@ -41,7 +41,7 @@ namespace MyFood.Models
         public string note { get; set; }
 
         [ForeignKey("FoodReceiptForm3")]
-        public int? f3_id { get; set; }
+        public long? f3_id { get; set; }
         public FoodReceiptForm3 FoodReceiptForm3 { get; set; }
 
     }
